requirements.txt
pytest>=7.0.0
pytest-html>=3.2.0
pytest-cov>=4.0.0
allure-pytest>=2.13.0
requests>=2.28.0
pyyaml>=6.0
python-dotenv>=1.0.0
